refactor: 안쓰는 파일 삭제 및 폴더 정리
