+++ week3/card_game/src/App.jsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from "styled-components";
 import theme from "./styles/theme";
 import { GlobalStyle } from "./styles/GlobalStyle";
-import Home from "./components/Main";
+import Home from "./components/Home";
 
 function App() {
   return (
